Always pre-compute operations
